refactor: extract correlated value util
